Fix namespace collision: move GeoJsonFieldService to Services.GeoJson
Services.Field namespace conflicted with Models.Field class, causing
CS0118 errors. Renamed to Services.GeoJson and fixed test assertion
for NumSections round-trip.
