+++ Tests/AgValoniaGPS.Services.Tests/GeoJsonFieldServiceTests.cs
@@ -1,6 +1,6 @@
 using AgValoniaGPS.Models;
 using AgValoniaGPS.Models.Base;
-using AgValoniaGPS.Services.Field;
+using AgValoniaGPS.Services.GeoJson;
 
 using MTrack = AgValoniaGPS.Models.Track.Track;
 using MTrackType = AgValoniaGPS.Models.Track.TrackType;
